feat: inventoryChangeNotifier 기능 구현 (매장에 넘겨줄 event 수정)
