+++ eatngo-core/src/main/kotlin/com/eatngo/inventory/service/InventoryService.kt
@@ -9,11 +9,7 @@ import com.eatngo.inventory.event.InventorySyncPublisher
 import com.eatngo.inventory.infra.InventoryPersistence
 import com.eatngo.product.dto.ProductCurrentStockDto
 import com.eatngo.product.dto.ProductDto
-import com.eatngo.product.service.StoreTotalInventoryTypeDecider
-import com.eatngo.inventory.event.InventoryEventPublisher
-import com.eatngo.inventory.event.InventoryChangedEvent
-import com.eatngo.inventory.event.InventoryChangedType
-import com.eatngo.inventory.event.InventoryChangedType.*
+import com.eatngo.product.service.InventoryChangeNotifier
 import org.springframework.cache.annotation.CacheEvict
 import org.springframework.cache.annotation.CachePut
 import org.springframework.cache.annotation.Cacheable
@@ -23,9 +19,8 @@ import org.springframework.transaction.annotation.Transactional
 @Service
 class InventoryService(
     private val inventoryPersistence: InventoryPersistence,
-    private val storeTotalInventoryTypeDecider: StoreTotalInventoryTypeDecider,
+    private val inventoryChangeNotifier: InventoryChangeNotifier,
     private val inventorySyncPublisher: InventorySyncPublisher,
-    private val inventoryEventPublisher: InventoryEventPublisher
 ) {
 
     @CachePut("inventory", key = "#productDto.id")
@@ -62,8 +57,9 @@ class InventoryService(
         val changedInventory = inventory.changeStock(productCurrentStockDto.action, productCurrentStockDto.amount)
         val savedInventory: Inventory = inventoryPersistence.save(changedInventory)
 
-        val eventType = storeTotalInventoryTypeDecider.decideInventoryType(
+        inventoryChangeNotifier.notifyInventoryStatusChange(
             storeId = storeId,
+            productId = productCurrentStockDto.id,
             initialStock = initialStock,
         )
 
@@ -73,15 +69,6 @@ class InventoryService(
                 syncedStock = savedInventory.stock
             )
         )
-
-        if (shouldPublishStoreStatusChangeEvent(eventType, initialStock)) {
-            inventoryEventPublisher.publishInventoryChangedEvent(
-                InventoryChangedEvent(
-                    productId = savedInventory.productId,
-                    inventoryChangedType = eventType
-                )
-            )
-        }
 
         return InventoryDto(savedInventory.quantity, savedInventory.stock)
     }
@@ -99,8 +86,9 @@ class InventoryService(
 
         val savedInventory = inventoryPersistence.save(inventory)
 
-        val eventType = storeTotalInventoryTypeDecider.decideInventoryType(
+        inventoryChangeNotifier.notifyInventoryStatusChange(
             storeId = productDto.storeId,
+            productId = savedInventory.productId,
             initialStock = initialStock
         )
 
@@ -111,27 +99,7 @@ class InventoryService(
             )
         )
 
-        if (shouldPublishStoreStatusChangeEvent(eventType, initialStock)) {
-            inventoryEventPublisher.publishInventoryChangedEvent(
-                InventoryChangedEvent(
-                    productId = savedInventory.productId,
-                    inventoryChangedType = eventType
-                )
-            )
-        }
-
         return InventoryDto(savedInventory.quantity, savedInventory.stock)
-    }
-
-    private fun shouldPublishStoreStatusChangeEvent(
-        eventType: InventoryChangedType, 
-        initialStock: Int
-    ): Boolean {
-        return when (eventType) {
-            OUT_OF_STOCK -> true
-            RESTOCKED -> initialStock == 0
-            else -> false
-        }
     }
 
 }
