+++ eatngo-core/src/main/kotlin/com/eatngo/product/event/ProductEventListener.kt
@@ -2,18 +2,14 @@ package com.eatngo.product.event
 
 import com.eatngo.common.exception.product.ProductException.ProductNotFound
 import com.eatngo.extension.orThrow
-import com.eatngo.inventory.event.InventoryChangedEvent
-import com.eatngo.inventory.event.InventoryChangedType.ADEQUATE_STOCK
-import com.eatngo.inventory.event.InventoryChangedType.RESTOCKED
-import com.eatngo.inventory.event.InventoryEventPublisher
 import com.eatngo.inventory.infra.InventoryCachePersistence
 import com.eatngo.order.domain.OrderItem
 import com.eatngo.order.event.OrderCanceledEvent
 import com.eatngo.order.event.OrderCreatedEvent
 import com.eatngo.order.event.OrderEvent
 import com.eatngo.product.infra.ProductPersistence
+import com.eatngo.product.service.InventoryChangeNotifier
 import com.eatngo.product.service.ProductService
-import com.eatngo.product.service.StoreTotalInventoryTypeDecider
 import org.springframework.context.event.EventListener
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
@@ -22,9 +18,8 @@ import org.springframework.transaction.annotation.Transactional
 @Transactional
 class ProductEventListener(
     private val inventoryCachePersistence: InventoryCachePersistence,
-    private val inventoryEventPublisher: InventoryEventPublisher,
     private val productPersistence: ProductPersistence,
-    private val storeTotalInventoryTypeDecider: StoreTotalInventoryTypeDecider,
+    private val inventoryChangeNotifier: InventoryChangeNotifier,
     private val productService: ProductService
 ) {
 
@@ -52,21 +47,11 @@ class ProductEventListener(
             .storeId
         val initialStock = productService.findTotalInitialStocks(storeId)
 
-        when (
-            val eventType = storeTotalInventoryTypeDecider.decideInventoryType(
-                storeId = storeId,
-                initialStock = initialStock
-            )
-        ) {
-            ADEQUATE_STOCK -> return
-            RESTOCKED -> return
-            else -> inventoryEventPublisher.publishInventoryChangedEvent(
-                InventoryChangedEvent(
-                    productId = orderItem.productId,
-                    inventoryChangedType = eventType
-                )
-            )
-        }
+        inventoryChangeNotifier.notifyInventoryStatusChange(
+            storeId = storeId,
+            productId = orderItem.productId,
+            initialStock = initialStock
+        )
     }
 
 }
